requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
beautifulsoup4==4.11.1
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
channels==3.0.3
channels-redis==4.0.0
chardet==4.0.0
charset-normalizer==3.0.1
constantly==15.1.0
cryptography==39.0.0
daphne==3.0.2
defusedxml==0.7.1
dj-database-url==1.2.0
Django==3.2.16
django-bootstrap-v5==1.0.11
django-bootstrap4==2.3.1
django-crispy-forms==1.14.0
django-on-heroku==1.1.2
django-redis==5.2.0
djangorestframework==3.12.4
firebase-admin==4.4.0
google-api-core==1.34.0
google-api-python-client==2.74.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.9.1
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
grpcio==1.51.1
grpcio-status==1.48.2
gunicorn==20.0.4
httplib2==0.21.0
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
msgpack==1.0.4
ndg-httpsclient==0.5.1
oauthlib==3.2.2
paypalrestsdk==1.13.1
Pillow==9.4.0
proto-plus==1.22.2
protobuf==3.20.3
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.7.1
redis==4.4.2
redis-simple-cache==0.0.8
requests==2.25.1
requests-oauthlib==1.3.1
rsa==4.9
service-identity==21.1.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stripe==2.55.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.14
whitenoise==6.3.0
zope.interface==5.5.2
